+++ Packages/ai.muna.muna/Runtime/Beta/BetaClient.cs
@@ -1,6 +1,6 @@
 /* 
 *   Muna
-*   Copyright © 2025 NatML Inc. All rights reserved.
+*   Copyright © 2026 NatML Inc. All rights reserved.
 */
 
 #nullable enable
@@ -10,8 +10,6 @@ namespace Muna.Beta {
     using API;
     using OpenAI;
     using Services;
-    using PredictorService = global::Muna.Services.PredictorService;
-    using EdgePredictionService = global::Muna.Services.PredictionService;
 
     /// <summary>
     /// Client for incubating features.
@@ -19,11 +17,6 @@ namespace Muna.Beta {
     public sealed class BetaClient {
 
         #region --Client API--
-        /// <summary>
-        /// Make predictions.
-        /// </summary>
-        public readonly PredictionService Predictions;
-
         /// <summary>
         /// OpenAI client.
         /// </summary>
@@ -36,10 +29,9 @@ namespace Muna.Beta {
         internal BetaClient(
             MunaClient client,
             PredictorService predictors,
-            EdgePredictionService predictions
+            PredictionService predictions
         ) {
-            this.Predictions = new PredictionService(client);
-            this.OpenAI = new OpenAIClient(predictors, predictions, Predictions.Remote);
+            this.OpenAI = new OpenAIClient(predictors, predictions);
         }
         #endregion
     }